requirements.txt
numpy==1.24.2
pandas==1.5.3
scikit-learn
matplotlib==3.7.0
seaborn==0.12.2
fastapi[all]==0.92.0
uvicorn==0.20.0
python-multipart==0.0.5
pydantic==1.10.5
pytest==7.2.1
pycodestyle==2.10.0
python-keycloak==2.14.0
fastapi_keycloak
